Ajout de la possibilité de rejoindre un channel public ou protégé par mot de passe (front only). Amélioration légère dans la liste des channels. Préparation rapide aux invitations de channels
